pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "perceptra-inference"
version = "0.1.0"
description = "Production-grade inference service for YOLO / RT-DETR ONNX models"
authors = [{name = "Tannous Geagea", email = "tannousgeagea@hotmail.com"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}

dependencies = [
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "opencv-python-headless>=4.8.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "typing-extensions>=4.8.0",
]

[project.optional-dependencies]
onnx = [
    "onnxruntime>=1.16.0",
]
onnx-gpu = [
    "onnxruntime-gpu>=1.16.0",
]
server = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "prometheus-client>=0.19.0",
    "httpx>=0.25.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]
all = [
    "perceptra-inference[onnx-gpu,server]",
]

[tool.setuptools.packages.find]
include = ["perceptra_inference*", "service*"]

[tool.ruff]
line-length = 100
